messages backend front end done -ws left

+++ api/server/api/team-member.js
@@ -3,6 +3,7 @@ const Team = require('../models/Team');
 const Channel = require('../models/Channel');
 const Invitation = require('../models/Invitation');
 const User = require('../models/User');
+const Message = require('../models/Message');
 const { signRequestForUpload } = require('../aws');
 
 const router = express();
@@ -13,9 +14,10 @@ router.post('/get-initial-data', async (req, res, next) => {
     let pendingAcceptances = [];
     let pendingInvitations = [];
     let currentUsers = [];
+    let messages = [];
     let result;
     const teams = await Team.getList({ userId: req.body.userId });
-    console.log(teams);
+
     if (teams.length > 0) {
       result = await Promise.all([
         Channel.getList({ teamId: teams[0]._id }),
@@ -30,6 +32,13 @@ router.post('/get-initial-data', async (req, res, next) => {
       pendingAcceptances = result[1];
       pendingInvitations = result[2];
       currentUsers = result[3];
+      if (channels.length > 0) {
+        messages = await Message.find({ channelId: channels[0]._id }).sort(
+          'createdAt'
+        );
+      }
+    } else {
+      pendingInvitations = await Invitation.find({ userId: req.body.userId });
     }
 
     return res.status(200).json({
@@ -38,7 +47,21 @@ router.post('/get-initial-data', async (req, res, next) => {
       pendingAcceptances,
       pendingInvitations,
       currentUsers,
+      messages,
     });
+  } catch (err) {
+    next(err);
+  }
+});
+
+router.post('/get-team-members', async (req, res, next) => {
+  try {
+    const team = await Team.findById(req.body.teamId);
+    const currentUsers = await User.find(
+      { _id: { $in: team.memberIds } },
+      '_id displayName email avatarUrl'
+    );
+    res.status(200).json({ currentUsers });
   } catch (err) {
     next(err);
   }
@@ -125,6 +148,12 @@ router.post('/add-team', async (req, res, next) => {
 
 router.post('/add-channel', async (req, res, next) => {
   try {
+    const team = await Team.findById(req.body.teamId);
+
+    if (!team.teamLeaderId.equals(req.userId))
+      return res.status(400).json({
+        err: 'You are not the owner of the team this channel belongs to.',
+      });
     const channel = new Channel({
       name: req.body.name,
       teamId: req.body.teamId,
@@ -138,8 +167,12 @@ router.post('/add-channel', async (req, res, next) => {
 
 router.post('/get-channels', async (req, res, next) => {
   try {
+    let messages = [];
     const channels = await Channel.getList({ teamId: req.body.teamId });
-    res.status(200).json({ channels });
+    if (channels.length > 0) {
+      messages = await Message.find({ channelId: channels[0]._id });
+    }
+    res.status(200).json({ channels, messages });
   } catch (err) {
     next(err);
   }
@@ -147,7 +180,13 @@ router.post('/get-channels', async (req, res, next) => {
 
 router.post('/delete-team', async (req, res, next) => {
   try {
-    await Team.deleteOne({ _id: req.body.teamId });
+    const team = await Team.findById(req.body.teamId);
+    if (!team.teamLeaderId.equals(req.userId))
+      return res
+        .status(400)
+        .json({ err: 'You are not the owner of this team.' });
+
+    await team.remove();
     await Channel.deleteMany({ teamId: req.body.teamId });
     res.status(200).json({});
   } catch (err) {
@@ -157,7 +196,13 @@ router.post('/delete-team', async (req, res, next) => {
 
 router.post('/delete-channel', async (req, res, next) => {
   try {
-    await Channel.deleteOne({ _id: req.body.channelId });
+    const channel = await Channel.findById(req.body.channelId);
+    const team = await Team.findById(channel.teamId);
+    if (!team.teamLeaderId.equals(req.userId))
+      return res.status(400).json({
+        err: 'You are not the owner of the team this channel belongs to.',
+      });
+    await channel.remove();
     res.status(200).json({});
   } catch (err) {
     next(err);
@@ -181,5 +226,41 @@ router.post(
     }
   }
 );
+
+router.post('/get-messages', async (req, res, next) => {
+  try {
+    const messages = await Message.find({ channelId: req.body.channelId }).sort(
+      'createdAt'
+    );
+    res.status(200).json({ messages });
+  } catch (err) {
+    next(err);
+  }
+});
+
+router.post('/add-message', async (req, res, next) => {
+  try {
+    const {
+      userId,
+      userEmail,
+      userDisplayName,
+      channelId,
+      userAvatarUrl,
+      text,
+    } = req.body;
+    const message = new Message({
+      userAvatarUrl,
+      userId,
+      userEmail,
+      userDisplayName,
+      channelId,
+      text,
+    });
+    await message.save();
+    res.status(200).json({ message });
+  } catch (err) {
+    next(err);
+  }
+});
 
 module.exports = router;
